+++ core/dnsserver/server_grpc.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/coredns/coredns/pb"
 	"github.com/coredns/coredns/plugin/pkg/transport"
-	"github.com/coredns/coredns/plugin/pkg/watch"
 
 	"github.com/grpc-ecosystem/grpc-opentracing/go/otgrpc"
 	"github.com/miekg/dns"
@@ -24,7 +23,6 @@ type ServergRPC struct {
 	grpcServer *grpc.Server
 	listenAddr net.Addr
 	tlsConfig  *tls.Config
-	watch      watch.Watcher
 }
 
 // NewServergRPC returns a new CoreDNS GRPC server and compiles all plugin in to it.
@@ -41,7 +39,7 @@ func NewServergRPC(addr string, group []*Config) (*ServergRPC, error) {
 		tlsConfig = conf.TLSConfig
 	}
 
-	return &ServergRPC{Server: s, tlsConfig: tlsConfig, watch: watch.NewWatcher(watchables(s.zones))}, nil
+	return &ServergRPC{Server: s, tlsConfig: tlsConfig}, nil
 }
 
 // Serve implements caddy.TCPServer interface.
@@ -103,9 +101,6 @@ func (s *ServergRPC) OnStartupComplete() {
 func (s *ServergRPC) Stop() (err error) {
 	s.m.Lock()
 	defer s.m.Unlock()
-	if s.watch != nil {
-		s.watch.Stop()
-	}
 	if s.grpcServer != nil {
 		s.grpcServer.GracefulStop()
 	}
@@ -142,12 +137,6 @@ func (s *ServergRPC) Query(ctx context.Context, in *pb.DnsPacket) (*pb.DnsPacket
 	}
 
 	return &pb.DnsPacket{Msg: packed}, nil
-}
-
-// Watch is the entrypoint called by the gRPC layer when the user asks
-// to watch a query.
-func (s *ServergRPC) Watch(stream pb.DnsService_WatchServer) error {
-	return s.watch.Watch(stream)
 }
 
 // Shutdown stops the server (non gracefully).
