+++ pb/Makefile
@@ -6,6 +6,8 @@
 all: dns.pb.go
 
 dns.pb.go: dns.proto
-	protoc --go_out=plugins=grpc:. dns.proto && \
-        sed -e s?golang.org/x/net/context?context? < dns.pb.go > dns.pb.go.tmp && \
-        mv dns.pb.go.tmp dns.pb.go
+	protoc --go_out=plugins=grpc:. dns.proto
+
+.PHONY: clean
+clean:
+	rm dns.pb.go
