+++ plugin/federation/kubernetes_api_test.go
@@ -3,7 +3,6 @@ package federation
 import (
 	"github.com/coredns/coredns/plugin/kubernetes"
 	"github.com/coredns/coredns/plugin/kubernetes/object"
-	"github.com/coredns/coredns/plugin/pkg/watch"
 
 	api "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,9 +18,6 @@ func (APIConnFederationTest) Stop() error                               { return
 func (APIConnFederationTest) SvcIndexReverse(string) []*object.Service  { return nil }
 func (APIConnFederationTest) EpIndexReverse(string) []*object.Endpoints { return nil }
 func (APIConnFederationTest) Modified() int64                           { return 0 }
-func (APIConnFederationTest) SetWatchChan(watch.Chan)                   {}
-func (APIConnFederationTest) Watch(string) error                        { return nil }
-func (APIConnFederationTest) StopWatching(string)                       {}
 
 func (APIConnFederationTest) PodIndex(string) []*object.Pod {
 	return []*object.Pod{
